Add x in Grid
